move cut params class to mirror slice hierarchy
